requirements.txt
numpy>=1.18.5,<1.25.0
scipy>=1.4.1
faiss-cpu>=1.7.0  # Use faiss-gpu for GPU support
flask>=2.0.0
pyyaml>=5.1
pytest>=6.2.5
requests>=2.25.0
scikit-learn>=0.24.2  # For LSH implementation and PCA
docker>=5.0.3
python-dotenv>=0.15.0
tqdm>=4.62.3
torch>=1.9.0  # For ResNet models
torchvision>=0.10.0  # For image processing and models
matplotlib>=3.4.0  # For visualization
pandas>=1.3.0
seaborn>=0.11.2
Pillow>=8.2.0  # For image processing
fpdf>=1.7.2  # For PDF report generation
